refactor metadata into callback
